Move the MinPlus_* Enums one package up
As the ConPwAffine algebra needs to take convolution from the pw_affine
and deconvolution from the con_pw_affine package, it should be defined
in their common superpackage (although Java does not really have a
notion of such hierarchical relations of packages).
For consistency, the same was done for the affine MinPlus enum.
