+++ src/main/java/org/networkcalculus/dnc/algebra/disco/MinPlus_Disco_ConPwAffine.java
@@ -23,19 +23,21 @@
  *
  */
 
-package org.networkcalculus.dnc.algebra.disco.affine;
+package org.networkcalculus.dnc.algebra.disco;
 
 import java.util.Set;
 
 import org.networkcalculus.dnc.algebra.MinPlus;
+import org.networkcalculus.dnc.algebra.disco.con_pw_affine.Deconvolution_Disco_ConPwAffine;
+import org.networkcalculus.dnc.algebra.disco.pw_affine.Convolution_Disco_PwAffine;
 import org.networkcalculus.dnc.curves.ArrivalCurve;
 import org.networkcalculus.dnc.curves.Curve;
 import org.networkcalculus.dnc.curves.MaxServiceCurve;
 import org.networkcalculus.dnc.curves.ServiceCurve;
 
-public enum MinPlus_Disco_Affine implements MinPlus {
-	MINPLUS_DISCO_AFFINE;
-  
+public enum MinPlus_Disco_ConPwAffine implements MinPlus {
+	MINPLUS_DISCO_CONPWAFFINE;
+	
 	// --------------------------------------------------------------------------------------------------------------
 	// Min-Plus-Operation Dispatching
 	// --------------------------------------------------------------------------------------------------------------
@@ -46,11 +48,11 @@ public enum MinPlus_Disco_Affine implements MinPlus {
 
 	// Service Curves
 	/* (non-Javadoc)
-	 * @see org.networkcalculus.dnc.algebra.MinPlus#convolve(org.networkcalculus.dnc.curves.ServiceCurve, org.networkcalculus.dnc.curves.ServiceCurve)
+	 * @see org.networkcalculus.dnc.algebra.IMinPlus#convolve(org.networkcalculus.dnc.curves.ServiceCurve, org.networkcalculus.dnc.curves.ServiceCurve)
 	 */
 	@Override
 	public ServiceCurve convolve(ServiceCurve service_curve_1, ServiceCurve service_curve_2) throws Exception {
-		return Convolution_Disco_Affine.convolve(service_curve_1, service_curve_2);
+		return Convolution_Disco_PwAffine.convolve(service_curve_1, service_curve_2);
 	}
 
 	/* (non-Javadoc)
@@ -59,7 +61,7 @@ public enum MinPlus_Disco_Affine implements MinPlus {
 	@Override
 	public Set<ServiceCurve> convolve(Set<ServiceCurve> service_curves_1,
 			Set<ServiceCurve> service_curves_2) throws Exception {
-		return Convolution_Disco_Affine.convolve(service_curves_1, service_curves_2);
+		return Convolution_Disco_PwAffine.convolve(service_curves_1, service_curves_2);
 	}
 
 	// Arrival Curves
@@ -68,7 +70,7 @@ public enum MinPlus_Disco_Affine implements MinPlus {
 	 */
 	@Override
 	public ArrivalCurve convolve(ArrivalCurve arrival_curve_1, ArrivalCurve arrival_curve_2) throws Exception {
-		return Convolution_Disco_Affine.convolve(arrival_curve_1, arrival_curve_2);
+		return Convolution_Disco_PwAffine.convolve(arrival_curve_1, arrival_curve_2);
 	}
 
 	/* (non-Javadoc)
@@ -76,7 +78,7 @@ public enum MinPlus_Disco_Affine implements MinPlus {
 	 */
 	@Override
 	public ArrivalCurve convolve(Set<ArrivalCurve> arrival_curves) throws Exception {
-		return Convolution_Disco_Affine.convolve(arrival_curves);
+		return Convolution_Disco_PwAffine.convolve(arrival_curves);
 	}
 
 	// Maximum Service Curves
@@ -86,7 +88,7 @@ public enum MinPlus_Disco_Affine implements MinPlus {
 	@Override
 	public MaxServiceCurve convolve(MaxServiceCurve max_service_curve_1, MaxServiceCurve max_service_curve_2)
 			throws Exception {
-		return Convolution_Disco_Affine.convolve(max_service_curve_1, max_service_curve_2);
+		return Convolution_Disco_PwAffine.convolve(max_service_curve_1, max_service_curve_2);
 	}
 
 	// Arrival Curves and Max Service Curves
@@ -96,7 +98,7 @@ public enum MinPlus_Disco_Affine implements MinPlus {
 	@Override
 	public Set<Curve> convolve_ACs_MaxSC(Set<ArrivalCurve> arrival_curves,
 			MaxServiceCurve maximum_service_curve) throws Exception {
-		return Convolution_Disco_Affine.convolve_ACs_MSC(arrival_curves, maximum_service_curve);
+		return Convolution_Disco_PwAffine.convolve_ACs_MSC(arrival_curves, maximum_service_curve);
 	}
 
 	/* (non-Javadoc)
@@ -105,7 +107,7 @@ public enum MinPlus_Disco_Affine implements MinPlus {
 	@Override
 	public Set<ArrivalCurve> convolve_ACs_MaxScRate(Set<ArrivalCurve> arrival_curves,
 			MaxServiceCurve extra_gamma_curve) throws Exception {
-		return Convolution_Disco_Affine.convolve_ACs_EGamma(arrival_curves, extra_gamma_curve);
+		return Convolution_Disco_PwAffine.convolve_ACs_EGamma(arrival_curves, extra_gamma_curve);
 	}
 
 	// ------------------------------------------------------------
@@ -117,7 +119,7 @@ public enum MinPlus_Disco_Affine implements MinPlus {
 	@Override
 	public Set<ArrivalCurve> deconvolve(Set<ArrivalCurve> arrival_curves, ServiceCurve service_curve)
 			throws Exception {
-		return Deconvolution_Disco_Affine.deconvolve(arrival_curves, service_curve);
+		return Deconvolution_Disco_ConPwAffine.deconvolve(arrival_curves, service_curve);
 	}
 
 	/* (non-Javadoc)
@@ -126,7 +128,7 @@ public enum MinPlus_Disco_Affine implements MinPlus {
 	@Override
 	public Set<ArrivalCurve> deconvolve(Set<ArrivalCurve> arrival_curves, Set<ServiceCurve> service_curves)
 			throws Exception {
-		return Deconvolution_Disco_Affine.deconvolve(arrival_curves, service_curves);
+		return Deconvolution_Disco_ConPwAffine.deconvolve(arrival_curves, service_curves);
 	}
 
 	/* (non-Javadoc)
@@ -143,7 +145,7 @@ public enum MinPlus_Disco_Affine implements MinPlus {
 	@Override
 	public ArrivalCurve deconvolve(ArrivalCurve arrival_curve, ServiceCurve service_curve,
 			boolean tb_rl_optimized) throws Exception {
-		return Deconvolution_Disco_Affine.deconvolve(arrival_curve, service_curve);
+		return Deconvolution_Disco_ConPwAffine.deconvolve(arrival_curve, service_curve);
 	}
 
 	/* (non-Javadoc)
@@ -152,6 +154,7 @@ public enum MinPlus_Disco_Affine implements MinPlus {
 	@Override
 	public Set<ArrivalCurve> deconvolve_almostConcCs_SCs(Set<Curve> curves,
 			Set<ServiceCurve> service_curves) throws Exception {
-		return Deconvolution_Disco_Affine.deconvolve_almostConcCs_SCs(curves, service_curves);
+		return Deconvolution_Disco_ConPwAffine.deconvolve_almostConcCs_SCs(curves, service_curves);
 	}
+	
 }
